Retain original waypoint in Hold, and restore when converting back
Previously, it was impossible to convert a Hold back into the original
Waypt after converting it into a Hold. With this change, the original
waypoint is retained in the Hold data structure, and we can dig it back
out in order to restore it when the hold counter is set to 0. This
allows aircraft systems to remove previously defined holds from the
flightplan.